Split memory search runtime boundary
